Add MachineRequestInterface::getType()
