+++ src/Model/MachineRequestInterface.php
@@ -4,6 +4,10 @@ namespace App\Model;
 
 interface MachineRequestInterface
 {
+    /**
+     * @return MachineProviderActionInterface::ACTION_*
+     */
+    public function getType(): string;
     public function getMachineId(): string;
     public function getRetryCount(): int;
     public function incrementRetryCount(): MachineRequestInterface;
