+++ tests/Functional/Controller/MachineControllerTest.php
@@ -8,6 +8,7 @@ use App\Controller\MachineController;
 use App\Entity\Machine;
 use App\Message\MachineRequestMessage;
 use App\Model\Machine\State;
+use App\Model\MachineProviderActionInterface;
 use App\Model\MachineRequest;
 use App\Model\ProviderInterface;
 use App\Repository\MachineRepository;
@@ -57,7 +58,10 @@ class MachineControllerTest extends AbstractBaseFunctionalTest
 
         $this->messengerAsserter->assertQueueCount(1);
 
-        $expectedRequest = new MachineRequest((string) $machine);
+        $expectedRequest = new MachineRequest(
+            MachineProviderActionInterface::ACTION_CREATE,
+            (string) $machine
+        );
         $expectedMessage = MachineRequestMessage::createCreate($expectedRequest);
         self::assertGreaterThan(0, $expectedRequest->getMachineId());
         $this->messengerAsserter->assertMessageAtPositionEquals(0, $expectedMessage);
