+++ tests/Functional/MessageDispatcher/MachineRequestMessageDispatcherTest.php
@@ -7,6 +7,7 @@ namespace App\Tests\Functional\MessageDispatcher;
 use App\Entity\Machine;
 use App\Message\MachineRequestMessage;
 use App\MessageDispatcher\MachineRequestMessageDispatcher;
+use App\Model\MachineProviderActionInterface;
 use App\Model\MachineRequest;
 use App\Model\ProviderInterface;
 use App\Services\MachineFactory;
@@ -55,7 +56,11 @@ class MachineRequestMessageDispatcherTest extends AbstractBaseFunctionalTest
 
         $this->defaultDispatcher->dispatch(
             MachineRequestMessage::createCreate(
-                new MachineRequest((string) $this->machine, 0)
+                new MachineRequest(
+                    MachineProviderActionInterface::ACTION_CREATE,
+                    (string) $this->machine,
+                    0
+                )
             )
         );
 
@@ -63,7 +68,11 @@ class MachineRequestMessageDispatcherTest extends AbstractBaseFunctionalTest
         $this->messengerAsserter->assertMessageAtPositionEquals(
             0,
             MachineRequestMessage::createCreate(
-                new MachineRequest((string) $this->machine, 0)
+                new MachineRequest(
+                    MachineProviderActionInterface::ACTION_CREATE,
+                    (string) $this->machine,
+                    0
+                )
             )
         );
 
@@ -79,7 +88,11 @@ class MachineRequestMessageDispatcherTest extends AbstractBaseFunctionalTest
 
         $this->updateMachineMessageDispatcher->dispatch(
             MachineRequestMessage::createGet(
-                new MachineRequest((string) $this->machine, 0)
+                new MachineRequest(
+                    MachineProviderActionInterface::ACTION_GET,
+                    (string) $this->machine,
+                    0
+                )
             )
         );
 
@@ -87,7 +100,11 @@ class MachineRequestMessageDispatcherTest extends AbstractBaseFunctionalTest
         $this->messengerAsserter->assertMessageAtPositionEquals(
             0,
             MachineRequestMessage::createGet(
-                new MachineRequest((string) $this->machine, 0)
+                new MachineRequest(
+                    MachineProviderActionInterface::ACTION_GET,
+                    (string) $this->machine,
+                    0
+                )
             )
         );
