+++ tests/Functional/MessageHandler/MachineRequestMessageHandlerTest.php
@@ -32,7 +32,10 @@ class MachineRequestMessageHandlerTest extends AbstractBaseFunctionalTest
 
     public function testInvokeForCreateAction(): void
     {
-        $request = new MachineRequest(md5('id content'));
+        $request = new MachineRequest(
+            MachineProviderActionInterface::ACTION_CREATE,
+            md5('id content')
+        );
         $message = MachineRequestMessage::createCreate($request);
 
         $createMachineHandler = (new MockCreateMachineHandler())
@@ -47,7 +50,10 @@ class MachineRequestMessageHandlerTest extends AbstractBaseFunctionalTest
 
     public function testInvokeForGetAction(): void
     {
-        $request = new MachineRequest(md5('id content'));
+        $request = new MachineRequest(
+            MachineProviderActionInterface::ACTION_GET,
+            md5('id content')
+        );
         $message = MachineRequestMessage::createGet($request);
 
         $createMachineHandler = (new MockCreateMachineHandler())
