+++ tests/Functional/Services/MachineHandler/UpdateMachineHandlerTest.php
@@ -75,7 +75,7 @@ class UpdateMachineHandlerTest extends AbstractBaseFunctionalTest
      * @dataProvider handleDataProvider
      *
      * @param array<ResponseInterface|\Throwable> $httpFixtures
-     * @param \App\Model\Machine\State::VALUE_* $currentState
+     * @param State::VALUE_* $currentState
      */
     public function testHandle(
         array $httpFixtures,
@@ -94,7 +94,11 @@ class UpdateMachineHandlerTest extends AbstractBaseFunctionalTest
         $this->machine->setState($currentState);
         $this->machineStore->store($this->machine);
 
-        $request = new MachineRequest((string) $this->machine, 0);
+        $request = new MachineRequest(
+            MachineProviderActionInterface::ACTION_GET,
+            (string) $this->machine,
+            0
+        );
         $outcome = $this->handler->handle($request);
 
         self::assertEquals($expectedOutcome, $outcome);
@@ -198,7 +202,11 @@ class UpdateMachineHandlerTest extends AbstractBaseFunctionalTest
 
         $this->setExceptionLoggerOnHandler($exceptionLogger);
 
-        $request = new MachineRequest((string) $this->machine, 0);
+        $request = new MachineRequest(
+            MachineProviderActionInterface::ACTION_GET,
+            (string) $this->machine,
+            0
+        );
         $outcome = $this->handler->handle($request);
 
         self::assertEquals(
@@ -229,7 +237,11 @@ class UpdateMachineHandlerTest extends AbstractBaseFunctionalTest
 
         $this->setExceptionLoggerOnHandler($exceptionLogger);
 
-        $request = new MachineRequest((string) $this->machine, $retryCount);
+        $request = new MachineRequest(
+            MachineProviderActionInterface::ACTION_GET,
+            (string) $this->machine,
+            $retryCount
+        );
         $outcome = $this->handler->handle($request);
 
         self::assertEquals(
@@ -266,7 +278,11 @@ class UpdateMachineHandlerTest extends AbstractBaseFunctionalTest
 
         $this->setExceptionLoggerOnHandler($exceptionLogger);
 
-        $request = new MachineRequest((string) $this->machine, 0);
+        $request = new MachineRequest(
+            MachineProviderActionInterface::ACTION_GET,
+            (string) $this->machine,
+            0
+        );
         $outcome = $this->handler->handle($request);
 
         self::assertEquals(
@@ -279,7 +295,11 @@ class UpdateMachineHandlerTest extends AbstractBaseFunctionalTest
     {
         $this->mockHandler->append(new Response(404));
 
-        $request = new MachineRequest((string) $this->machine, 11);
+        $request = new MachineRequest(
+            MachineProviderActionInterface::ACTION_GET,
+            (string) $this->machine,
+            11
+        );
         $outcome = $this->handler->handle($request);
 
         self::assertEquals(
